Code/sqa-system/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SQA-System"
version = "0.1.0"
description = "Master Thesis implementation."
authors = [
  { name = "Marco Schneider" }
]
license = { text = "MIT" }
dependencies = [
  "langchain~=0.3.18",
  "langchain-chroma~=0.2.2",
  "langchain-community~=0.3.17",
  "langchain-core~=0.3.35",
  "langchain-huggingface~=0.1.2",
  "langchain-google-genai~=2.0.9",
  "langchain-openai~=0.3.5",
  "langchain-text-splitters~=0.3.6",
  "langchain-ollama~=0.2.3",
  "ollama~=0.4.7",
  "pandas~=2.2.3",
  "nltk~=3.9.1",
  "ragas~=0.2.14",
  "numpy~=1.26.4",
  "sentence-transformers~=3.2.0", # Requirements for Huggingface models
  "inquirerpy~=0.3.4",
  "pytest~=8.3.3",
  "typer<10.0.0", # Needs to be below 10.0.0 for compatibility chromadb ops
  "cryptography~=43.0.1",
  "orkg~=0.21.3",
  "seaborn~=0.13.2",
  "matplotlib~=3.9.2",
  "weave~=0.51.37",
  "tqdm~=4.67.1",
  "sparqlwrapper~=2.0.0",
  "pybtex~=0.24.0",
  "language_tool_python~=2.8.1",
  "ipywidgets~=8.1.5",
  "pylatexenc~=2.10",
  "sacrebleu~=2.5.1",
  "rouge_score~=0.1.2",
  "lightrag-hku==1.1.5",
  "bitsandbytes~=0.45.0",
  "accelerate~=1.2.1",
  "evaluate~=0.4.3",
  "bert_score~=0.3.13", # Requirement for the Bert-Score-Evaluator
  "sentencepiece", # requirement for huggingface models
  "aioboto3~=13.3.0", # requirement for lightrag-hku
  "nano_vectordb", # requirement for lightrag-hku
  "pipmaster", # requirement for lightrag-hku
  "rank-bm25~=0.2.2", # requirement for Think-on-Graph
  "chromadb-ops~=0.1.3", # Tool for hnsw index maintanace
]

[project.optional-dependencies]
# Only install graphrag for Python versions below 3.13
graphrag = [
  "graphrag==1.2.0; python_version < '3.13'"
]
# Only install codecarbon when NOT installing graphrag
codecarbon = [
  "codecarbon~=2.8.3",
]

[tool.setuptools.packages.find]
where = ["."]
